requirements.txt
Flask
Flask-SQLAlchemy
mysql-connector-python
mysqlclient
openai
python-dotenv
